requirements.txt
cmake
dlib
mediapipe
ffmpeg
ffmpeg-python
imutils
opencv-python
Pillow
scikit-image
scipy
openai
streamlit